Address review: close multi-header leak path + 0600 settings.local.json
Two Yellow findings from the PR #403 review (CROW-402):

Y1 — Multi-header launch line no longer leaves a ~/.zshrc leak. When a gateway
has more than one header the value can't go on the launch line (embedded
newline), so it was carried only by settings.local.json — but the prefix was a
bare `ANTHROPIC_BASE_URL=… claude`, which doesn't clear an inherited
ANTHROPIC_CUSTOM_HEADERS. The multi-header branch now emits
`unset ANTHROPIC_CUSTOM_HEADERS && ANTHROPIC_BASE_URL=… ` in both
ClaudeLaunchArgs.gatewayEnvPrefix and setup.sh's gateway_launch_prefix, so the
gateway's baseURL is never paired with stale global headers. Tests updated.

Y2 — Resolved secrets in settings.local.json are now written 0600. The env block
caches a post-`op read` bearer token, so HookConfigGenerator.writeGatewayEnv and
setup.sh's write_settings_local now chmod the file owner-only, matching
ConfigStore's 0600 on config.json. Added a permission assertion to the bash test.

Also softened the docs / Settings UI copy: `op://` means "kept out of
config.json," not "no secret on disk" — the resolved value is cached (0600) in
the worktree's settings.local.json. Added a precedence note to configuration.md
(G1): the launch-line assignment is what reliably overrides a shell export; the
end state is to delete the global ~/.zshrc exports.


Crow-Session: E3DFAF9C-38CB-4D95-979D-9CC48F4EC3C9
